requirements.txt
pyyaml
dictdiffer